refactor(analysis-types): split entropy DTOs
