+++ crates/tokmd-analysis-types/src/lib.rs
@@ -18,6 +18,7 @@ mod api_surface;
 mod derived;
 mod duplication;
 mod effort;
+mod entropy;
 pub mod findings;
 mod git;
 mod supply;
@@ -44,6 +45,7 @@ pub use effort::{
     EffortDeltaClassification, EffortDeltaReport, EffortDriver, EffortDriverDirection,
     EffortEstimateReport, EffortModel, EffortResults, EffortSizeBasis, EffortTagSizeRow,
 };
+pub use entropy::{EntropyClass, EntropyFinding, EntropyReport};
 pub use git::{
     BusFactorRow, ChurnTrend, CodeAgeBucket, CodeAgeDistributionReport, CommitIntentCounts,
     CommitIntentKind, CommitIntentReport, CorporateFingerprint, CouplingRow, DomainStat,
@@ -129,33 +131,6 @@ pub struct AnalysisArgsMeta {
 pub struct Archetype {
     pub kind: String,
     pub evidence: Vec<String>,
-}
-
-// -----------------
-// Entropy profiling
-// -----------------
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct EntropyReport {
-    pub suspects: Vec<EntropyFinding>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct EntropyFinding {
-    pub path: String,
-    pub module: String,
-    pub entropy_bits_per_byte: f32,
-    pub sample_bytes: u32,
-    pub class: EntropyClass,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum EntropyClass {
-    Low,
-    Normal,
-    Suspicious,
-    High,
 }
 
 // -------------
